backend/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.10.0
bcrypt==5.0.0
billiard==4.2.2
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distlib==0.4.0
distro==1.9.0
dj-database-url==3.1.0
Django==5.2.10
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
ecdsa==0.19.1
fastapi==0.127.0
filelock==3.20.0
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.12.0
kombu==5.5.4
mysql-connector-python==9.5.0
numpy==2.4.0
openai==2.14.0
packaging==25.0
passlib==1.7.4
pgvector==0.4.2
pillow==12.0.0
pipenv==2025.0.4
platformdirs==4.5.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlparse==0.5.3
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.40.0
vine==5.1.0
virtualenv==20.35.4
wcwidth==0.2.14
whitenoise==6.11.0
